system scheduler: account for alloc resources when checking node feasibility
CE part of nomad-enterprise PR
https://github.com/hashicorp/nomad-enterprise/pull/3372
